+++ src/client/RootAction.ts
@@ -11,13 +11,7 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-import {
-  Update,
-  Recipe,
-  StatusOr,
-  Food,
-  ConversionData,
-} from "../core";
+import { Update, Recipe, StatusOr, Food, ConversionData } from "../core";
 import { Config } from "./config";
 
 export enum ActionType {
