+++ src/core/Food.ts
@@ -17,7 +17,7 @@ import { Quantity } from "./Quantity";
 
 /**
  * Representation of a food used for recipe calculations.
- * 
+ *
  * This data can be derived from various sources, e.g. FDC Data
  * or other recipes.
  */
